Styles on Special:upload page
Workaround to present extension stylesheet in Special:Upload page, which can't be modified. This is the fastest possible way. 
If the case is "upload", a line is added in the injected form field, embedding the stylesheet in an inline style@scoped element. This attribute isn't standard, though, and UAs don't actually implement it. But localised CSS rules apply only to <div#categoryselectmaster>, so this doesn't generate an issue.